refactor: move subscription operator rules to dedicated file
Move SubscriptionOperatorRule base class and all subscription-based
operator health validation rules to a new dedicated file
